Cast BFS face swap transformer to requested dtype after LoRA fusion
The Qwen-Image-Edit model stores weights in bfloat16. After fusing LoRA
weights, explicitly cast the transformer to the requested dtype (float32
by default) to avoid mat1/mat2 dtype mismatch during forward pass.

+++ bfs_face_swap/pytorch/loader.py
@@ -112,7 +112,7 @@ class ModelLoader(ForgeModel):
         )
         self.pipeline.fuse_lora()
 
-        self._transformer = self.pipeline.transformer
+        self._transformer = self.pipeline.transformer.to(dtype)
         self._transformer.eval()
         return self._transformer
